// TypeScript/src/models/enums/ItemType.ts
export enum ItemType {
    GOOD_WINE = "Good Wine",
    BACKSTAGE_PASSES_REFACTOR = "Backstage passes for Re:Factor",
    BACKSTAGE_PASSES_HAXX = "Backstage passes for HAXX",
    B_DAWG_KEYCHAIN = "B-DAWG Keychain",
    DUPLICATE_CODE= "Duplicate Code",
    LONG_METHODS = "Long Methods",
    UGLY_VARIABLE_NAMES = "Ugly Variable Names"
}